rewrite: move functions for updating refs to `MutableRepo`
The functions now depend only on `MutableRepo`, so I think they belong
on that type. This gets us closer to being able to make
`parent_mapping` private again.
